requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface_hub
faiss-cpu
pypdf
sentence-transformers

